Add files via upload
This is the created notebook for the Python code to process files generated when implementing the 'Contacts' tool in ChimeraX 1.7 for proteins. It takes 2 files, the one that serves as reference and the query one. The code parses these documents and counts the number of common contacts for both protein structures. 
** It is recommended that the contacts files are checked beforehand, since sometimes they vary in the number of columns, thus you may require to adjust the index numbers to the ones corresponding to the last three columns for both "atom 1" and "atom 2". Sample reference (AF_contacts) and query files (Galaxy4_0_contacts) are provided in this repository.